Split to sub-clients
