+++ crates/bitwarden-organization-invite-link/src/lib.rs
@@ -1,8 +1,14 @@
 #![doc = include_str!("../README.md")]
 
+mod error;
+mod invite_link_admin_client;
 mod invite_link_client;
+mod invite_link_user_client;
 mod organization_invite_link;
-pub use invite_link_client::{InviteLinkClient, InviteLinkClientExt, InviteLinkError};
+pub use error::InviteLinkError;
+pub use invite_link_admin_client::InviteLinkAdminClient;
+pub use invite_link_client::{InviteLinkClient, InviteLinkClientExt};
+pub use invite_link_user_client::InviteLinkUserClient;
 pub use organization_invite_link::{
     OrganizationInviteLink, OrganizationInviteLinkSsoView, OrganizationInviteLinkStatusView,
     OrganizationInviteLinkView,
